// models/User.ts
export interface User {
    id: number;
    firstName: string;
    lastName: string;
    maidenName: string;
    username: string;
    password: string;
    birthDate: string;
    age: number;
    gender: string;
    email: string;
    phone: string;
    image: string;
    bloodGroup: string;
    height: number;
    weight: number;
    eyeColor: string;
  }